==> Instacart basket analysis/03. Scripts/1.1 Data import and descriptive analysis.ipynb <==
# 1.1 Data import and descriptive analysis
** **
## Table of contents:

 1. Importing libraries
 2. Importing dataframe <br>
 3. Descriptive analysis with basic functions<br>
 4. Creating a list to import specific columns in a dataframe
** **

# 1. Importing libraries
** **

# Importing libraries
import pandas as pd
import numpy as np
import os

# 2. Importing dataframe
** **

In this section we will import our first dataframe: the orders dataframe.

# Importing the dataframe
df = pd.read_csv(r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis\02. Data\Original data\orders.csv', index_col = False)

# Creating a path variabile for the folder
path = r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis'

# Testing the path created
path

# Importing again the dataframe, but with the path method
df = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'orders.csv'), index_col = False)

# Checking the output with the head function
df.head ()

# 3. Descriptive analysis with basic functions
** **

In this section we will perform an initial descriptive analysis using basic functions.

# Using the head function with a custom argument
df.head(30)

The output is showing the first 30 rows, because of the custom argument.

# Using the tail() function to check last 5 rows of the dataframe
df.tail()

# Checking the shape of the dataframe, with the shape function
df.shape

The shape function indicates the total number of rows and columns of the dataframe.

# Using the info() function
df.info()

The info function gives informations about columns names and respective datatypes.

# Using the dtypes function
df.dtypes

The dtypes function shows the datatype of all the dataframe columns.

# Using the columns function
df.columns

The columns function shows all columns names.

# Using the describe() function
df.describe()

The describe function gives an overview of the descriptive statistics of the columns: <br>
Count of values, mean, standard deviations, min and max values, and 1st-2nd-3rd quartiles.

# 4. Creating a list to import specific columns in a dataframe
** **

In this section we create a custom list of variables before importing the dataframe.

# Creating a list containing only specific columns
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

In the list, the eval_set column is not included

# Importing the dataframe, using the list as an additional argument
df = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# Checking the head
df.head()

The eval_set column was filtered out of when the dataframe was imported.
