==> Instacart basket analysis/03. Scripts/1.2 Data import and descriptive analysis.ipynb <==
# 1.2 Data import and descriptive analysis
** **
## Table of contents:

 1. Importing libraries
 2. Importing first dataframe <br>
 3. Importing second dataframe <br>
 4. Descriptive analysis on the second dataframe
** **

# 1. Importing libraries
** **

import pandas as pd
import numpy as np
import os

# 2. Importing first dataframe
** **

# Creating a path variabile for the folder
path = r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis'

# Testing the path
path

# Creating a list of columns to be included in the dataframe
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Testing the list contents
vars_list

# Importing the dataframe without index column and with the new list created
df = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'orders.csv'), index_col = False, usecols = vars_list)

# Printing the first 5 rows
df.head()

# 3. Importing second dataframe
** **

In this section we will import our second dataframe: the products dataframe.

df_prods = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'products.csv'), index_col = False)

# 4. Descriptive analysis on the second dataframe
** **

# Printing top 20 rows
df_prods.head(20)

# Printing last 35 rows
df_prods.tail(35)

# Printing the name of the columns
df_prods.columns

# Checking the shape
df_prods.shape

# Checking the maximum value of aisle_id
df_prods.describe()

Maximum value for aisle_id is 134.

# Checking the datatype of the department_id column
df_prods.dtypes

The datatype of the department_id columun is int64 (integer).
