==> Instacart basket analysis/03. Scripts/2. Data wrangling.ipynb <==
# 2. Data wrangling
** **
## Table of contents:

1. Importing libraries
2. Importing the dataframes <br>
3. Data wrangling
    - 3.1 Removing a column <br>
    - 3.2 Searching for missing values <br>
    - 3.3 Calculating the frequency
    - 3.4 Renaming a column <br>
    - 3.5 Changing datatypes <br>
    - 3.6 Transposing <br>
        - 3.6.1 Fixing the header issue <br>
    - 3.7 Creating a subset <br>
4. Tasks
    - 4.1 Task 2
    - 4.2 Task 3
    - 4.3 Task 4
    - 4.4 Task 5
    - 4.5 Task 6
    - 4.6 Task 7
    - 4.7 Task 8
    - 4.8 Task 9
    - 4.9 Task 10
    - 4.10 Exporting the dataframes
** **

# 1. Importing libraries
** **

import pandas as pd
import numpy as np
import os

# 2. Importing the dataframes
** **

# Creating a path variabile for the folder
path = r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis'

# Importing dataframe without index column
df_ords = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'orders.csv'), index_col = False)

# Printing the first 5 rows to test
df_ords.head()

# Importing second dataframe
df_prods = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'products.csv'), index_col = False)

# Printing the first 5 rows to test
df_prods.head()

# 3. Data wrangling
** **

## 3.1 Removing a column

# Removing a column (only visually) with df.drop
df_ords.drop(columns = ['eval_set'])

## 3.2 Searching for missing values

# Searching for missing values (NaN)
df_ords['days_since_prior_order'].value_counts(dropna = False)

## 3.3 Calculating the frequency

# Calculating the frequency of values in a specific column
df_ords['days_since_prior_order'].value_counts()

## 3.4 Renaming a column

# Renaming a column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Testing changing name
df_ords.head()

## 3.5 Changing datatype

# Changing datatype into string
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Checking the change
df_ords['order_id'].dtype

"O" means "object, the way Pandas call strings.

## 3.6 Transposing

# Importing a third dataframe
df_dep = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'departments.csv'), index_col = False)

# Checking the head
df_dep.head()

# Transposing a dataframe
df_dep.T

# Creating a new dataframe transposed
df_dep_t = df_dep.T

# Testing the output of the new dataframe
df_dep_t

### 3.6.1 Fixing the header issue

# Creating an index with the reset_index() function
df_dep_t.reset_index()

# creating a variable containing the headers on row 0
new_header = df_dep_t.iloc[0]

# Testing the variable created
new_header

# Creating a new dataframe containing all the rows except row 0 (the old header)
df_dep_t_new = df_dep_t[1:]

# Testing the new dataframe created
df_dep_t_new

# Changing the name of the columns, using the variable created earlier
df_dep_t_new.columns = new_header

# Testing
df_dep_t_new

The variable created became the header of the new dataframe.

# Creating a variable that contains the dataframe as a data dictionary
data_dict = df_dep_t_new.to_dict('index')

# Testing the output
data_dict

# Looking at department_id variable in the dataframe
df_prods.head()

# Printing the department_id 19 using the data dictionary
print(data_dict.get('19'))

## 3.7 Creating a subset

# Creating a new dataframe, filtering data on a specific criteria
df_snacks = df_prods[df_prods['department_id']==19]

We created a new dataframe containing only the rows where department_id is 19. But how does it works?

# Testing the part of code that checks for value 19.
df_prods['department_id']==19

Basically Python search in every row if there is a 19 in that specific column. If Yes, returns True.
If we enclose this part of code into brackets, it can visually show some of the rows containing the filtered value.

df_prods[df_prods['department_id']==19]

This is only a visual results. To "save" it, you have to redefine (or create) a new dataframe, as done in the first line of this section.

# Checking the new data frame created
df_snacks.head()

#Testing other two methods. First one, using loc function:
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

#testing new dataframe created
df_snacks_2

# Testing second method, using loc and isin:
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

# Testing new dataframe created
df_snacks_3

# 4. Tasks

# Checking variables of df_ords
df_ords.head()

# Checking datatypes
df_ords.dtypes

## 4.1 Task 2

User_id is a list of numbers used as identifiers. For this reason, it should be stored as strings.

# Changing variable to a string format
df_ords['user_id'] = df_ords['user_id'].astype('str')

# Testing the change
df_ords.dtypes

The datatype has been successfully changed.

## 4.2 Task 3

# Changing the name of a column without overwriting the dataframe
df_ords.rename(columns = {'order_hour_of_day' : 'order_hour_of_creation'}, inplace = True)

# Testing the change
df_ords

## 4.3 Task 4

In this section we need to identify the hour with most orders created.

# Looking for the busiest hour. Checking frequency.
df_ords['order_hour_of_creation'].value_counts()

The hour with most orders created is 10.

## 4.4 Task 5

In this section we need to find out which department has 4 as a department_id.

# Looking for value 4 in department_id using a data dictionary
print(data_dict.get('4'))

# Looking at the complete dictionary
data_dict

## 4.5 Task 6

In this section we need to create a subset containing only breakfast items. Breakfast department_id is 14.

# Creating the subset as a new dataframe
df_prods_breakfast = df_prods[df_prods['department_id']==14]

# Testing the output
df_prods_breakfast

## 4.6 Task 7

In this section we will create a subset using multiple values from the department_ids.

# Creating a subset with multiple value from department_id: 5 (alcohol), 7 (beverages), 12 (meat seafood) and 20 (deli).
df_prods_dinner = df_prods.loc[df_prods['department_id'].isin([5, 7, 12, 20])]

# Testing the output
df_prods_dinner

# Testing again, checking more rows
df_prods_dinner.head (30)

The dataframe contains only rows with the specified department_id.

## 4.7 Task 8

# Checking the shape of the dinner dataframe
df_prods_dinner.shape

This new dataframe has 7650 rows. I got the same result during the output in task 7.

## 4.8 Task 9

In this section we perform some checks around a suspect user, with user_id 1.

#Creating a subset to check the strange behaviour of user_id 1.
df_suspect_user = df_ords[df_ords['user_id'] == '1']

# Testing the output
df_suspect_user

# Checking the head
df_suspect_user.head (20)

This user placed only 11 orders. The user placed orders only from Sunday (1) to Wednesday (4) and from hours 7 to 16. <br>
It made also order 8 and 9 on the same day, because order 9 has 0 days since prior order.

## 4.9 Task 10

In this section we try to find more info about the suspect behaviour, using descriptive statistics.

# Checking descriptive statistics
df_suspect_user.describe()

There isn't much additional information. <br>
This user orders on average on Monday, at 10, and places a new order on average every 19 days. <br>
The min and max values for days_since_prior_order is 0 and 30.

## 4.10 Exporting the dataframes

In this sections we will export as csv files the two dataframes wrangled. <br>
Before exporting the orders dataframe, I still need to drop the "eval_set" column.

# Removing the eval_set column
df_ords = df_ords.drop(columns = ['eval_set'])

# Checking the change
df_ords

# Exporting df_ords
df_ords.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'orders_wrangled.csv'))

# Exporting df_dep_t_new
df_dep_t_new.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'departments_wrangled.csv'))
