==> Instacart basket analysis/03. Scripts/3. Data consistency checks.ipynb <==
# 3. Data consistency checks
** **
## Table of contents:

1. Importing libraries <br>
2. Importing dataframes <br>
3. Data consistency checks
    - 3.1 Mixed-type data <br>
    - 3.2 Missing values <br>
    - 3.3 Duplicates <br>
    - 3.4 Exporting the dataframe after consistency checks <br>
4. Tasks
    - 4.1 Task 2
    - 4.2 Task 3
    - 4.3 Task 5
    - 4.4 Task 6
    - 4.5 Task 7
    - 4.6 Exporting dataframes after consistency checks
** **

# 1. Importing libraries
** **

import pandas as pd
import numpy as np
import os

# 2. Importing dataframes
** **

# Creating a path variabile for the folder
path = r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis'

# Importing orders wrangled dataframe from Prepared Data
df_ords = pd.read_csv(os.path.join(path, '02. Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# Printing the first 5 rows to test
df_ords.head()

There is a strange column, like some sort of fake index.

# Removing first column
df_ords = df_ords.drop(columns = ['Unnamed: 0'])

# Testing
df_ords

# Importing second dataframe
df_prods = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'products.csv'), index_col = False)

# Printing the first 5 rows to test
df_prods.head()

# 3. Data consistency checks
** **

## 3.1 Mixed-type data

In this section is shown how to check if a column contains mixed-type data.

# Creating a new dataframe for test
df_test = pd.DataFrame ()

# Creating a column with mixed data
df_test['mix'] = ['a', 'b', 1, True]

This column contains mixed-type data: two strings, an integer and a boolean value.

# Testing
df_test.head()

# Checking for mixed-type columns
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

This code checks every column in the specified dataframe (df_test in this case) and prints the name of the columns containing mixed-type data (mix). <br>
If nothing is printed, it means the dataframe doesn't contain mixed-type columns.

## 3.2 Missing values

In this section we see how is possible to check for missing values and how to deal with them.

# Checking for missing values
df_prods.isnull().sum()

The column "product_name" contains 16 missing values.

# Subsetting (creating a new dataframe) to visualize the missing values
df_nan = df_prods[df_prods['product_name'].isnull() == True]

# Visualizing subset
df_nan.head(20)

Here we can see a subset containing only the 16 rows with missing values.

# Checking the shape of the dataframe before deleting missing values
df_prods.shape

# Creating a new dataframe that excludes the 16 rows with missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Checking if new dataset has 16 rows less...
df_prods_clean.shape

The new dataframe df_prods_clean has 16 rows less. We successfully filtered out the missing values.

## 3.3 Duplicates

In this section we see how to find duplicates and remove them.

# Creating a new dataframe containing only duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

# Checking the new dataframe
df_dups

# Checking the number of rows before deleting duplicates
df_prods_clean.shape

# Removing duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# Checking the number of orws after deletion
df_prods_clean_no_dups.shape

The new dataframe has 5 rows less. Duplicates have been removed.

## 3.4 Exporting the dataframe after consistency checks

# Exporting the new dataframe, that does not contain missing values and duplicates
df_prods_clean_no_dups.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'products_checked.csv'))

# 4. Tasks
** **

In this section, we will perform consistency checks on the orders dataframe. <br>
Before diving into the tasks, I noticed that the datatype for two variables in ords dataframe is wrong. I will fix them.

# Changing the datatypes
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords['user_id'] = df_ords['user_id'].astype('str')

## 4.1 Task 2

In this section we will use the describe function to see if we find anything strange in the products dataframe.

# Using describe on df_prods
df_prods_clean_no_dups.describe()

Max value for prices is 99999 this is extremely suspicious, considering that the mean is just 9. <br>
It should be an outlier made by mistake (or a missing value, with 99999 acting as a placeholder). <br>
Note: the issue has been addressed in a future script, 4.9.

# Creating a new subset to inspect the strange value
df_prices_sus = df_prods[df_prods['prices']==99999]

# Checking the output
df_prices_sus

It seems there is a single product (2 % Reduced Fat Milk) with this wrong price.

## 4.2 Task 3

In this section we will check if there are mixed-typde data columns in the orders dataframe.

# Checking mixed data-type in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)


<b> Observations: </b> <br>
No column printed. It means there aren't variables with mixed data-type in df_ords.

## 4.3 Task 5

In this section we will check if there are missing values inside the columns of the orders dataframe.

# Checking for missing values in df_ords
df_ords.isnull().sum()

# Investigating the missing values through a new subset
df_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

# Checking the output
df_nan

<b> Observations: </b> <br>
days_since_prior_order has 206209 missing values. <br>
My guess is that this is not a mistake, neither a missing value. <br>
The column contains a "missing value" everytime a new customer places the first order (see how the user_id changes, while the order_number is always 1). <br> 
Being the first order, is not possible to calculate how many days since the prior order, because there isn't a prior order.

## 4.4 Task 6

# Addressing the missing values using an appropriate method
df_ords_clean = df_ords[df_ords['days_since_prior_order'].isnull() == False]

I decided to create a new dataframe that filtered out the missing values. <br>
It doesn't make sense to impute values in this case. <br>
It's the first order, so is not possible to establish how much days since the prior order, because this "order" does not exists. <br>
Doing so, we are losing information on 206.209 records, but compared to the whole dataframe is only the 6%.

# Testing the new dataframe, that does not contains missing values
df_ords_clean

The new dataframe has 206.209 rows less.

## 4.5 Task 7

In this section, we will check for duplicates in the orders dataframe.

# Checking for duplicates in the new clean dataframe
df_ords_dups = df_ords_clean[df_ords_clean.duplicated()]

# Checking the new dataframe
df_ords_dups

The new dataframe contains 0 rows. 0 duplicates found.

# As a bonus, I would like to check if there were duplicates in the dataframe prior to deleting the rows with missing values
df_ords_dups_bonus = df_ords[df_ords.duplicated()]

# Checking the output
df_ords_dups_bonus

No duplicates found even prior to missing values deletion.

## 4.6 Exporting dataframes after consistency checks

In this sections we will export both orders and products dataframes after consistency checks. <br>
Before exporting them, I will perform some additional checks to be sure everything is in order. <br>
Note: the orders dataframe exported will be df_ords_clean because we didn't find any duplicates neither mixed-type data.

# Checking head and tail
df_ords_clean

<b> Observations: </b> <br>
The dataframe now has 3.214.874 rows (206.209 rows with missing values have been deleted), however the index is the same as before and does not match with the new number of rows. <br>
I will check with tutor to see if I should reset the index. <br>
NOTE: When imported in script 4.6, the dataframe index reset on its own and the issue is fixed.

# Checking descriptive statistics
df_ords_clean.describe()

<b> Observations: </b> <br>
Having deleted all the rows with missing values in days_since_prior_order column (and subsequently, all rows with 1 as order number), the minimum value for order number is now 2.

# Checking dtypes
df_ords_clean.dtypes

<b> Observations: </b> <br>
Everything seems correct. I'm in doubt if order_number should be also converted in string. <br>
NOTE: the datatype has been changed in a successive script.

I will perform the same checks also on the products dataframe.

# Checking head and tail
df_prods_clean_no_dups

# Checking describe
df_prods_clean_no_dups.describe()

<b> Observations: </b> <br>
Here it is again the matter of the max value of prices, that is going to be investigated in the future.

# Checking dtypes
df_prods_clean_no_dups.dtypes

<b> Observations: </b> <br>
I don't know if product_id, aisle_id and department_id should be strings as well. I will check with my tutor. <br>
NOTE: the datatype has been changed in a successive script.

Both databases are now ready to be exported.

# Exporting the new orders dataframe that does not contain missing values
df_ords_clean.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'orders_checked.csv'))

# Exporting the new products dataframe that does not contain missing values and duplicates
df_prods_clean_no_dups.to_csv(os.path.join(path, '02. Data', 'Prepared Data', 'products_checked.csv'))
