==> Instacart basket analysis/03. Scripts/4.1 Merging orders and orders products prior dataframes.ipynb <==
# 4.1 Merging "orders" and "orders products prior" dataframes
** **
## Table of contents:

1. Importing libraries <br>
2. Importing dataframes <br>
3. Merging the two dataframes
4. Exporting the new dataframe in a pickle format
** **

# 1. Importing libraries
** **

import pandas as pd
import numpy as np
import os

# 2. Importing dataframes
** **

# Creating a path variabile for the folder
path = r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis'

# Importing orders dataframe from Prepared Data
df_ords = pd.read_csv(os.path.join(path, '02. Data', 'Prepared Data', 'orders_checked.csv'), index_col= False)

# Importing order products prior dataframe from Original Data
df_ords_prior = pd.read_csv(os.path.join(path, '02. Data', 'Original Data', 'order_products_prior.csv'), index_col = False)

# Checking the output of orders dataframe
df_ords.head()

# Removing the first column
df_ords = df_ords.drop(columns = ['Unnamed: 0'])

# Testing
df_ords.head()

# Checking the output of orders products prior dataframe
df_ords_prior.head()

# Checking the shape of both dataframes
df_ords.shape

df_ords_prior.shape

The new dataframe has only 4 columns, but 32+ millions rows.

# 3. Merging the two dataframes

The two dataframes have different lenghts, so the best way to combine them is with a merging. <br>
Both dataframes have the order_id columns, that can be used as a key. <br>
We can also create a flag column to see if the merge was successfully.

# Merging the two dataframes and adding an indicator column
df_merged_large = df_ords.merge(df_ords_prior, on = ['order_id'], indicator = True)

# Checking the head of the new dataframe
df_merged_large.head()

#Checking the shape
df_merged_large.shape

<b> Observations: </b> <br>
The merged dataframe has 30+ millions rows, and 4 new columns (product_id, add_to_cart_orde, reordered, merge). <br>
The new dataframe has less rows thant the original orders_products_prior because some order_id were missing from the order dataframe, due to the elimination of the rows with missing values. (Script 4.5)

# Checking the values of the indicator column
df_merged_large['_merge'].value_counts()

<b> Observations: </b> <br>
Only "both" values. It means there was a full match between the two dataframes during merge.

# 4. Exporting the new dataframe in a pickle format
** **

# Exporting new dataframe in pkl format
df_merged_large.to_pickle(os.path.join(path, '02. Data','Prepared Data', 'orders_products_combined.pkl'))
