==> Instacart basket analysis/03. Scripts/4.2 Merging orders products combined and products dataframes.ipynb <==
# 4.2 Merging "orders products combined" and "products" dataframes
** **
## Table of contents:

1. Importing libraries <br>
2. Importing dataframes <br>
3. Merging the two dataframes
4. Exporting the new dataframe in a pickle format
** **

# 1. Importing libraries
** **

import pandas as pd
import numpy as np
import os

# 2. Importing dataframes
** **

# Creating a path variabile for the folder
path = r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis'

# Importing the previously merged dataframe from Prepared Data
df_ords = pd.read_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'orders_products_combined.pkl'))

# Checking the shape of the merged dataframe
df_ords.shape

# Checking the head
df_ords.head()

# Removing the _merge column that is not necessary anymore.
df_ords = df_ords.drop(columns = ['_merge'])

# Testing the change
df_ords.head()

# Importing products dataframe
df_prods = pd.read_csv(os.path.join(path, '02. Data', 'Prepared Data', 'products_checked.csv'), index_col= False)

# Checking the head
df_prods.head()

# Removing the first column
df_prods = df_prods.drop(columns = ['Unnamed: 0'])

# Testing the change
df_prods.head()

# 3. Merging the two dataframes
** **

For this merge I want a full match, in order that my final dataframe contains only informations that are in both dataframes. <br>
For this reason I will use an inner join and I will create a merge flag. <br>
The key column (the column both dataframes share) is the product_id.

# Merging the two dataframes on the product_id column
df_merged = df_ords.merge(df_prods, on = ['product_id'], indicator = True)

# Checking the output
df_merged

<b> Observations: </b>
Previous orders dataframe had 30.356.421 rows and 9 columns. <br>
This new dataframe has 27.687 less rows (probably due to inner join, there were some rows that got excluded) and five more columns (product_name, aisle_id, department_id, prices and  the merge flag).

# Cheking the results of the merge through the merge flag
df_merged['_merge'].value_counts()

Of course, being an inner join, there are only full matchs, so all values are "both".

# Before exporting the dataframe, I am going to drop the _merge column.
df_merged = df_merged.drop(columns = ['_merge'])

# Testing the drop
df_merged.head()

# 4. Exporting the new dataframe in a pickle format
** **

Due to the size of the dataframe, is not possible to export it as a CVS. <br>
It will be exported in a PKL format.

# Exporting new dataframe in pkl format
df_merged.to_pickle(os.path.join(path, '02. Data','Prepared Data', 'orders_products_merged.pkl'))
