==> Instacart basket analysis/03. Scripts/5. Deriving new variables.ipynb <==
# 5. Deriving new variables
** **
## Table of contents:

1. Importing libraries <br>
2. Importing dataframe <br>
3. Deriving a new variable
    - 3.1 Creating an "indicator column" for "prices"
        - 3.1.1 Method number 1: defining a custom function
        - 3.1.2 Method number 2: using the .loc function
        - 3.1.3 Creating the new "price_range" column in the real dataframe
    - 3.2 Creating a new column to check the busiest day
        - 3.2.1 Creating the new column with a for-loop
4. Tasks
    - 4.1 Creating a new column to check the two most busiest and least busiest days
    - 4.2 Creating a new column to check the hours with most orders
    - 4.3 Exporting the new dataframe with derived variables
** **

# 1. Importing libraries
** **

import pandas as pd
import numpy as np
import os

# 2. Importing dataframe
** **

# Creating a path variabile for the folder
path = r'C:\Users\Simone\Desktop\Career Foundry\Esercizi modulo 5\Instacart basket analysis'

# Importing the merged dataframe from Prepared Data
df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02. Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Checking the head
df_ords_prods_merged.head()

# 3. Deriving a new variable
** **

In this section will be shown how is possible to derive (create) a new variable (column) from already existing variables.

## 3.1 Creating an "indicator column" for "prices"

First of all, we will test the process on a subset.

# Creating a subset to test (containing only 1 million rows)
df = df_ords_prods_merged[:1000000]

# Checking the shape of the subset
df.shape

### 3.1.1 Method number 1: defining a custom function

# Defining a new function
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

The following code defines a new function called "price_label" that will check every row of the dataframe and return the following values ('Low-range product', 'Mid-range product', 'High range') if the conditions on the prices column are met. <br>
Let's apply the function to our subset, using .apply.

# Using the new function on the subset to create the new column 'price_range'
df['price_range'] = df.apply(price_label, axis=1)

<b> Observations: </b> <br>
^ This is not an error message per se, but Python is informing us that this procedure could cause issue and recommends using the loc method. <br>
We can ignore this message. The function has been applied anyway.

# Checking the values in the new column
df['price_range'].value_counts(dropna = False)

There isn't a single product labeled as "High range". It does mean that in this subset, there isn't a single product costing more than 15€.

# Testing what is the max value for the prices column in the subset
df['prices'].max()

<b> Observations: </b> <br>
This confirms that there aren't products costing more than 15€ in the subset created.

### 3.1.2 Method number 2: using the .loc function

This second method uses the .loc function to establish the conditions of the new column.

# Creating first condition
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Creating second condition
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

# Creating third condition
df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Checking the values in the new column
df['price_range_loc'].value_counts(dropna = False)

<b> Observations: </b> <br>
The results are the same as the first method used.

### 3.1.3 Creating the new "price_range" column in the real dataframe

We don't need the subset anymore. It's the time to derive the new variable in the real dataframe, using the .loc method.

# Condition 1
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Condition 2
df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

# Condition 3
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Checking the values in the new column
df_ords_prods_merged['price_range_loc'].value_counts(dropna = False)

Here is the distribution of the products in the dataframe. It seems that more than 60% of products ordered is mid-range.

# Testing the head
df_ords_prods_merged.head()

## 3.2 Creating a new column to check the busiest day

In this section we will derive a new variable that will indicate us if the order has been placed in the most busy days or in the least busy.

### 3.2.1 Creating the new column with a for-loop

# Checking the frequency of the interested column
df_ords_prods_merged['orders_day_of_week'].value_counts (dropna = False)

<b> Observations: </b> <br>
According to the count of values, the day most busy (highest frequency) is Saturday, while the least busy is Wednesday.

# Creating a list that will check for the values and label them
result = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

# Creating a new column in the dataframe, containing the list created
df_ords_prods_merged['busiest_day'] = result

# Checking the head
df_ords_prods_merged.head()

#Counting the values
df_ords_prods_merged['busiest_day'].value_counts(dropna = False)

Over 60% of the orders are placed in days that are regularly busy, while almost 6 millions are placed during the busiest day (Saturday) and 3 millions and a half during the least busy (Wednesday).

# 4. Tasks
** **

Before diving to the tasks, I will change the name gof the first derived column, that is not intuitive enough.

# Renaming the price_range_loc column
df_ords_prods_merged.rename(columns = {'price_range_loc' : 'price_label'}, inplace = True)

# Testing the change
df_ords_prods_merged.head()

## 4.1 Creating a new column to check the two most busiest and least busiest days

We are going to create a new column that will flag the two most busy days and the two least busy days.

# Creating a new list that will check for the values and label them
result_2 = []

for value in df_ords_prods_merged["orders_day_of_week"]:
  if value == 0 or value == 1:
    result_2.append("Busiest day")
  elif value == 4 or value == 3:
    result_2.append("Least busy")
  else:
    result_2.append("Regularly busy")

0 and 1 (Saturday and Sunday) will be flagged as "Busiest day", while 3 and 4 (Tuesday and Wednesday) will be flagged as "Least busy".

# Creating a new column in the dataframe, containing the list created
df_ords_prods_merged['busiest_days'] = result_2

# Checking the head
df_ords_prods_merged.head()

The new column has been successfully added.

# Checking the values in the new column
df_ords_prods_merged['busiest_days'].value_counts(dropna = False)

<b> Observations: </b> <br>
Sum of values is 30.328.763 (as the total rows of the dataframe). No missing values. <br>
In this new column, the "regularly busy" days decreased from around 20 millions to around 12 millions, because two more days (Sunday and Tuesday) were "flagged" in a different category. <br>
"Busiest days" increased from almost 6 millions to 11 millions (+5) while "Least busy" days increased from around 3 millions and a half to 7. <br>
I think this new columns contains data that is more on point, because the values are better distribuited. <br>
Also, in the original column, the days 0 and 1 (and 3 and 4 as well) had a similar frequency, so it makes sense to aggregate these days together as the busiests and least busiests.

## 4.2 Creating a new column to check the hours with most orders

In order to derive a new column and flag the hours in some categories, we need to check hours frequency.

# Checking the frequency of the "order hour of creation" column
df_ords_prods_merged['order_hour_of_creation'].value_counts (dropna = False)

<b> Observations: </b> <br>
The hours should be divided and labeled as “Most orders,” “Average orders,” and “Fewest orders.” <br>
Looking at the frequency, the top 8 results goes from 9 to 16, and should be labeled as the hours with "Most orders". <br>
During the morning (7 and 8) and during 17 to 23, the number of orders slowly decreases, so I will label these hours as "Average orders". <br>
From 0 to 6 is the period with the "Fewest orders", so these hours will be labeled accordingly.

# Creating a new list that will check for the values and label them
result_3 = []

for value in df_ords_prods_merged["order_hour_of_creation"]:
  if value in [9, 10, 11, 12, 13, 14, 15, 16]:
    result_3.append("Most orders")
  elif value in [0, 1, 2, 3, 4, 5, 6]:
    result_3.append("Fewest orders")
  else:
    result_3.append("Average orders")

# Creating a new column in the dataframe, containing the list created
df_ords_prods_merged['busiest_period_of_day'] = result_3

# Testing
df_ords_prods_merged.head(10)

The new column has been added successfully.

# Checking the frequency of the new column
df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

<b> Observations: </b> <br>
Around 60% of the orders are placed during the hours with the "Most orders". <br>
Around 30% of the orders are placed during the hours with "Average orders". <br>
Only a small percentage of orders are placed from 0 to 6 ("Fewest orders).

## 4.3 Exporting the new dataframe with derived variables

# Exporting the new dataframe (with derived columns) in pkl format
df_ords_prods_merged.to_pickle(os.path.join(path, '02. Data','Prepared Data', 'orders_products_merged_derived.pkl'))
